pyproject.toml
[build-system]
requires = ["setuptools>=70", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "ccx-messaging"
dynamic = ["version"]
readme = "README.md"
authors = [
    {name="CCX Processing Team", email="ccx-core-processing-team@redhat.com"}
]
classifiers = [
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]

requires-python = ">= 3.8"
dependencies = [
    "app-common-python>=0.2.3",
    "boto3>=1.34.1,<1.42.22",  # limited due to incompatibility between s3fs and newer versions of boto
    "confluent-kafka>=2.0.0",
    "insights-core>=3.1.2",
    "insights-core-messaging>=1.2.15",
    "jsonschema>=4.0.0",
    "prometheus-client>=0.16.0",
    "python-json-logger>=2.0.7,<5",
    "PyYAML>=6.0",
    "requests>=2.31.0",
    "sentry-sdk>=1.37.1",
    "watchtower>=3.0.0"
]

[tool.setuptools.packages.find]
include = ["ccx_messaging"]

[project.license]
text = "Apache-2.0"

[project.urls]
Homepage = "https://github.com/RedHatInsights/insights-ccx-messaging"

[project.scripts]
ccx-messaging = "ccx_messaging.command_line:ccx_messaging"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "freezegun"
]
dev = [
    "coverage",
    "pycco",
    "ruff"
]

[tool.setuptools_scm]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "UP", "C", "D"]
ignore = ["D211", "C401", "D203", "D213", "UP006", "UP007", "UP009", "UP015", "UP035"]
